hope

+++ src/Controller/LuckyController.php
@@ -12,54 +12,12 @@ use App\Repository\UserRepository;
 use App\Repository\TaskRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Routing\Annotation\Route;
+use function Symfony\Component\String\s;
 
 class LuckyController extends AbstractController
 {
 
 
-    /**
-     * @Route("/user", name="registrate")
-     */
-    public function registration(Request $request, ValidatorInterface $validator, UserRepository $userRepository): Response
-    {
-        $entityManager = $this->getDoctrine()->getManager();
-
-        $name = $request->request->get('name');
-        $password = $request->request->get('password');
-        if($name==""||$password=="") return new Response('Введите корректные данные');
-        $user = new User();
-        $user->setName($name);
-        $user->setPassword($password);
-
-        $errors = $validator->validate($user);
-        $checkUser = $userRepository
-            ->findOneBy(['name' => $name]);
-        if(($checkUser!==null)&&($checkUser->getPassword()===$password)){
-            return new Response('С возвращением, '.$name);
-
-        }else{
-            if (count($errors) > 0) {
-                /*
-                 * Uses a __toString method on the $errors variable which is a
-                 * ConstraintViolationList object. This gives us a nice string
-                 * for debugging.
-                 */
-                $errorsString = (string) $errors;
-                return new Response($errorsString);
-            }
-            else{
-                $entityManager->persist($user);
-                $entityManager->flush();
-                $mresponse = array('name'=>$name, 'password'=>$password);
-                return new Response(
-                    json_encode($mresponse)
-                );
-            }
-        }
-
-
-
-    }
 
 
     /**
@@ -67,40 +25,49 @@ class LuckyController extends AbstractController
      */
     public function tasksList(Request $request, UserRepository $userRepository, TaskRepository $taskRepository): Response
     {
+
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        /** @var \App\Entity\User $user */
+        $user = $this->getUser();
+
         $entityManager = $this->getDoctrine()->getManager();
-
         if($request->getMethod() == 'GET') {
-            $name = $request->query->get('name');
-            $checkUser = $userRepository->findOneBy(['name' => $name]);
-            $tasks = $checkUser->getTasks();
-
-            foreach ($tasks as $task){
-                $mresponse[$task->getId()]=$task->getTaskText();
+            $tasks = $user->getTasks();
+            if(sizeof($tasks) == 0){
+                $mresponse["message"] = "There is no tasks";
+            } else{
+                foreach ($tasks as $task){
+                    $mresponse[$task->getId()]=$task->getTaskText();
+                }
             }
 
-
+            $mresponse["task_owner"]=$user;
             return new Response(
                 json_encode($mresponse)
             );
-        }else{
-            $name = $request->request->get('name');
-            $checkUser = $userRepository->findOneBy(['name' => $name]);
-
+        }else if($request->getMethod() == 'POST'){
 
             $taskText = $request->request->get('text');
 
             $task = new Task();
             $task->setTaskText($taskText);
 
-            $checkUser->addTask($task);
-            $entityManager->persist($task);
-            $entityManager->flush();
+            $user->addTask($task);
+            //$entityManager->persist($task);
+            //$entityManager->flush();
 
-            $mresponse = array('Owner'=>$name, 'Text'=>$taskText);
+            $mresponse["message"] = "Task added.";
+            $mresponse["task_text"] = $taskText;
+            $mresponse["task_owner"] = $user;
             return new Response(
                 json_encode($mresponse)
             );
 
+        } else{
+            $mresponse["message"] = "Wrong request type.";
+            return new Response(
+                json_encode($mresponse)
+            );
         }
     }
 
@@ -110,28 +77,40 @@ class LuckyController extends AbstractController
      */
     public function taskManagment(Request $request,TaskRepository $taskRepository, int $task_id): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        /** @var \App\Entity\User $user */
+        $user = $this->getUser();
+
         $entityManager = $this->getDoctrine()->getManager();
+
         $checkTask = $taskRepository->find($task_id);
-        if($request->getMethod() == 'DELETE'){
+        if($checkTask->getOwner() == $user){
+            if($request->getMethod() == 'DELETE'){
 
-            $entityManager->remove($checkTask);
-            $entityManager->flush();
-            return new Response(
-                'ваша задача удалена'
-            );
+                $entityManager->remove($checkTask);
+                $entityManager->flush();
 
-        }else{
-            $newText = $request->get('new_text');
-            $checkTask->setTaskText($newText);
-            $entityManager->flush();
+                $mresponse["message"] = "Task deleted.";
+                return new Response(
+                    json_encode($mresponse)
+                );
+
+            }else{
+                $newText = $request->get('new_text');
+                $checkTask->setTaskText($newText);
+                $entityManager->flush();
+                $mresponse["message"] = "Task updated.";
+                return new Response(
+                    json_encode($mresponse)
+                );
+            }
+        } else {
+            $mresponse["message"] = "Access denied";
             return new Response(
-                'ваша задача обновлена'
+                json_encode($mresponse)
             );
         }
 
-        return new Response(
-            'error'
-        );
     }
 
 }
